+++ src/main/java/org/kakara/engine/renderobjects/mesh/MeshType.java
@@ -1,0 +1,8 @@
+package org.kakara.engine.renderobjects.mesh;
+
+public enum MeshType {
+    SYNC,
+    ASYNC,
+    MULTITHREAD,
+    MODIFEDASYNC
+}
