+++ src/main/java/org/kakara/engine/renderobjects/mesh/RenderMesh.java
@@ -1,0 +1,7 @@
+package org.kakara.engine.renderobjects.mesh;
+
+public interface RenderMesh {
+    int getVertexCount();
+    void render();
+    void cleanUp();
+}
